+++ src/EventIDs.cpp
@@ -404,6 +404,7 @@ const wxWindowIDRef EventIDs::menu_add_path(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_evaluate_all_visible(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_evaluate_all(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_show_tip(wxWindow::NewControlId());
+const wxWindowIDRef EventIDs::menu_accessibility_help(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_show_cellbrackets(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_print_cellbrackets(wxWindow::NewControlId());
 const wxWindowIDRef EventIDs::menu_copy_matlab_from_worksheet(wxWindow::NewControlId());